taskforge_scaffold/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cerebras_cloud_sdk==1.29.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
packaging==25.0
pydantic==2.11.4
pydantic_core==2.33.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.1.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
webvtt-py==0.5.1
Werkzeug==3.1.3
zipp==3.21.0
